+++ src/validate-changelog.test.ts
@@ -531,6 +531,23 @@ describe('validateChangelog', () => {
       ).resolves.not.toThrow();
     });
 
+    it('should throw if the changelog has an empty release and noEmptyReleases is true', async () => {
+      const changelogWithEmptyRelease = changelogWithReleases.replace(
+        '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
+        '## [1.0.0] - 2020-01-01\n',
+      );
+      await expect(
+        validateChangelog({
+          changelogContent: changelogWithEmptyRelease,
+          currentVersion: '1.0.1',
+          repoUrl:
+            'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
+          isReleaseCandidate: false,
+          noEmptyReleases: true,
+        }),
+      ).rejects.toThrow("Release has no changelog entries: '1.0.0'");
+    });
+
     it('should throw if the changelog has an empty change category', async () => {
       const changelogWithEmptyChangeCategory = changelogWithReleases.replace(
         '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
@@ -623,6 +640,23 @@ describe('validateChangelog', () => {
           isReleaseCandidate: true,
         }),
       ).resolves.not.toThrow();
+    });
+
+    it('should throw if the changelog has an empty release and noEmptyReleases is true', async () => {
+      const changelogWithEmptyRelease = changelogWithReleases.replace(
+        '## [1.0.0] - 2020-01-01\n### Changed\n- Something else\n',
+        '## [1.0.0] - 2020-01-01\n',
+      );
+      await expect(
+        validateChangelog({
+          changelogContent: changelogWithEmptyRelease,
+          currentVersion: '1.0.0',
+          repoUrl:
+            'https://github.com/ExampleUsernameOrOrganization/ExampleRepository',
+          isReleaseCandidate: true,
+          noEmptyReleases: true,
+        }),
+      ).rejects.toThrow("Release has no changelog entries: '1.0.0'");
     });
 
     it('should throw if the changelog has an empty change category', async () => {